I've splitted out the Rabbit converter as a separte class. And I don't know why the method names are as `Saturnuni2zg` because it's the Rabbit Converter. Anyway, I thought it'd be nice to have meaningful method names. I renamed the names.
